agency-dashboard:  rework `useFetchTestConnection`
it's now called `useSyncJetpackConnectionStatus` and  not using on* handlers anymore as those are  removed from the RQ API  in v5
